修改pipeleine

+++ lad/lad/pipelines.py
@@ -57,25 +57,26 @@ class LadPipeline(object):
             if not os.path.exists(dir_path):
                 os.makedirs(dir_path)
 
-            for image_url in item['imageUrls']:
-                list_name = image_url.split('/')
-                file_name = list_name[len(list_name)-1]#图片名称
+            if item['imageUrls'] is not None:
+                for image_url in item['imageUrls']:
+                    list_name = image_url.split('/')
+                    file_name = list_name[len(list_name)-1]#图片名称
 
-                file_path = '%s/%s' % (dir_path, file_name)
+                    file_path = '%s/%s' % (dir_path, file_name)
 
-                if os.path.exists(file_name):
-                    continue
-                with open(file_path,'wb') as file_writer:
-                    conn = urllib.urlopen(image_url)#下载图片
-                    file_writer.write(conn.read())
+                    if os.path.exists(file_name):
+                        continue
+                    with open(file_path,'wb') as file_writer:
+                        conn = urllib.urlopen(image_url)#下载图片
+                        file_writer.write(conn.read())
 
-                # 七牛云上传图片
-                key = image_url.split('/')[-1]
-                new_url = self.qiniu_domain + key
-                new_imgarray.append(new_url)
-                token = self.qiniu_client.upload_token(self.qiniu_bucket_name, key)
-                put_file(token, key, file_path)
+                    # 七牛云上传图片
+                    key = image_url.split('/')[-1]
+                    new_url = self.qiniu_domain + key
+                    new_imgarray.append(new_url)
+                    token = self.qiniu_client.upload_token(self.qiniu_bucket_name, key)
+                    put_file(token, key, file_path)
 
-                os.remove(file_path)
+                    os.remove(file_path)
             item['imageUrls'] = new_imgarray
             return item
